split js files by features
