+++ index.html
@@ -55,7 +55,7 @@
 				<input id="french" class="language__radio" type="radio" name="language">
 				<input id="japanese" class="language__radio" type="radio" name="language">
 				<input id="korean" class="language__radio" type="radio" name="language">
-				<input id="chinese" class="language__radio" type="radio" name="language">
+				<input id="mandarin" class="language__radio" type="radio" name="language">
 				<!-- Tabs -->
 				<div class="language__tabs">
 					<div class="language__europe">
@@ -67,7 +67,7 @@
 					<div class="language__asia">
 						<label for="japanese" class="language__tab">Japanese</label>
 						<label for="korean" class="language__tab">Korean</label>
-						<label for="chinese" class="language__tab">Chinese</label>
+						<label for="mandarin" class="language__tab">Mandarin</label>
 					</div>
 				</div>
 
@@ -145,13 +145,13 @@
 					</div>
 
 					<!-- Chinese Content -->
-					<div class="language__content" data-tab="chinese">
+					<div class="language__content" data-tab="mandarin">
 						<form class="search__form">
 							<input class="search-form__input" type="text" placeholder="Artist or songs">
 							<label class="search-form__label">Search for artists or songs</label>
 							<button type="submit" class="search-form__btn">Search</button>
 						</form>
-						<div class="search-results" data-country="cn">
+						<div class="search-results" data-country="tw">
 							<!-- Chinese search results -->
 						</div>
 					</div>
@@ -164,7 +164,7 @@
 			<small class="footer__copyright">&copy; 2026 Global Music Finder</small>
 		</div>
 	</footer>
-	<script src="main.js"></script>
+	<script src="js/main.js" type="module"></script>
 </body>
 
 </html>
