+++ js/config.js
@@ -1,0 +1,54 @@
+// iTunes API
+export const API_BASE_URL = 'https://itunes.apple.com/search/';
+
+// List of featured artists by country
+export const artists = [
+	// English (US/UK artists, country code: gb)
+	{ artist: 'Adele', countryCode: 'gb' },
+	{ artist: 'Ed Sheeran', countryCode: 'gb' },
+	{ artist: 'Taylor Swift', countryCode: 'gb' },
+	{ artist: 'Beyoncé', countryCode: 'gb' },
+	{ artist: 'Harry Styles', countryCode: 'gb' },
+
+	// Spanish
+	{ artist: 'Rosalía', countryCode: 'es' },
+	{ artist: 'Bad Bunny', countryCode: 'es' },
+	{ artist: 'Alejandro Sanz', countryCode: 'es' },
+	{ artist: 'Pablo Alborán', countryCode: 'es' },
+	{ artist: 'Shakira', countryCode: 'es' },
+
+	// German
+	{ artist: 'Rammstein', countryCode: 'de' },
+	{ artist: 'Helene Fischer', countryCode: 'de' },
+	{ artist: 'Nena', countryCode: 'de' },
+	{ artist: 'Andreas Bourani', countryCode: 'de' },
+	{ artist: 'Herbert Grönemeyer', countryCode: 'de' },
+
+	// French
+	{ artist: 'Stromae', countryCode: 'fr' },
+	{ artist: 'Aya Nakamura', countryCode: 'fr' },
+	{ artist: 'Zaz', countryCode: 'fr' },
+	{ artist: 'Maître Gims', countryCode: 'fr' },
+	{ artist: 'Angèle', countryCode: 'fr' },
+
+	// Japanese
+	{ artist: '宇多田ヒカル', countryCode: 'jp' },
+	{ artist: 'YOASOBI', countryCode: 'jp' },
+	{ artist: 'あいみょん', countryCode: 'jp' },
+	{ artist: 'King Gnu', countryCode: 'jp' },
+	{ artist: 'Official髭男dism', countryCode: 'jp' },
+
+	// Korean
+	{ artist: 'BTS', countryCode: 'kr' },
+	{ artist: 'BLACKPINK', countryCode: 'kr' },
+	{ artist: 'IU', countryCode: 'kr' },
+	{ artist: 'Seventeen', countryCode: 'kr' },
+	{ artist: 'Stray Kids', countryCode: 'kr' },
+
+	// Chinese
+	{ artist: '周杰倫', countryCode: 'tw' },
+	{ artist: '五月天', countryCode: 'tw' },
+	{ artist: '蔡依林', countryCode: 'tw' },
+	{ artist: '林俊傑', countryCode: 'tw' },
+	{ artist: '陳奕迅', countryCode: 'tw' },
+];
